// src/interfaces/UserInterface.ts
export interface UsersResponse {
    page:        number;
    per_page:    number;
    total:       number;
    total_pages: number;
    data:        User[];
    support:     Support;
}

export interface User {
    id?:         number;
    email:      string;
    first_name: string;
    last_name:  string;
    avatar?:     string;
}

export interface UserList{
    page:       number;
    users:       User[]
    per_page:    number;
    total:       number;
    total_pages: number;
    support:     Support;
}

export interface Support {
    url:  string;
    text: string;
}

export interface UserResponse {
    data: User;
    support: Support;
}

export interface PostUserResponse{
    user: User
    status: number
}